+++ infrastructure/docker/services/router/generate-ssl.sh
@@ -4,10 +4,9 @@
 
 BASE=$(dirname $0)
 
-rm -rf mkdir $BASE/certs/
+CERTS_DIR=$BASE/certs
 
-CERTS_DIR=$BASE/etc/ssl/certs
-
+rm -rf $CERTS_DIR
 mkdir -p $CERTS_DIR
 
 openssl req -x509 -sha256 -newkey rsa:4096 \