+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Migrate from Invoke to Castor
 * Add a dockerfile linter
+* Do not store certificates in the router image
 
 ## 3.11.0 (2023-05-30)